+++ ReleaseNotes.md
@@ -6,6 +6,7 @@ A. Start with the README : https://github.com/SteveGilham/altcode.test/blob/mast
 * [BREAKING] make the `AssertionMatch` type specific to each package to avoid name clashes if multiple test suites are in use
 * [BREAKING] remove support for obsolescing Expecto `Expect.floatEqual`
 * [BREAKING] remove support for obsolescing NUnit `Assert.AreEqual<Nullable<double>>` methods
+* [BREAKING] rename `AltCode.Test.Xunit.AltAssert.DoesNotAssertionMatch` & `AssterionMatches` to `AltCode.Test.Xunit.AltAssert.DoesNotMatch` and `Matches` to line up with the methods being shadowed in `Xunit.Assert`
 * support `netstandard2.0` only (meaning ≥ `net472` compatible)
 * support Expecto ≥ 9.0.4, NUnit ≥ 3.13.3, Xunit ≥ 2.4.2
 * new for `Expecto.Expect`
@@ -26,6 +27,11 @@ A. Start with the README : https://github.com/SteveGilham/altcode.test/blob/mast
   * isLessThan
   * isLessThanOrEqual
 * new for `Expecto.CSharp.Function`, type `AltCSharpExpect` for the type-homogenous versions of `IsFasterThan`
+* new for `NUnit.Framework.Assert`
+  * Greater
+  * GreaterOrEqual
+  * Less
+  * LessOrEqual
 
 # 1.0.11
 * support Expecto 9.x